+++ CHANGELOG.md
@@ -8,6 +8,7 @@ format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Added
 - Many tests.
 - The function `is_full(grid)` returns true if the grid is full.
+- Subroutine `create_puzzle_with_unique_solution(grid, nb_empty)`
 
 ### Changed
 - Code improved, modernized and refactored.